add DELETE movie feature

+++ internal/data/movies.go
@@ -86,21 +86,46 @@ func (m MovieModel) Get(id int64) (*Movie, error) {
 
 // Update movie
 func (m MovieModel) Update(movie *Movie) error {
-	stmt := `
+	query := `
 		UPDATE movies
-		SET title = $1, year = $2, runtime = $3, version = $4
+		SET title = $1, year = $2, runtime = $3, genres = $4, version = version + 1
 		WHERE id = $5
+		RETURNING version
 	`
-	_, err := m.DB.Exec(stmt, movie.Title, movie.Year, movie.RunTime, movie.Version, movie.ID)
-	return err
+	args := []interface{}{
+		movie.Title,
+		movie.Year,
+		movie.RunTime,
+		pq.Array(movie.Genres),
+		movie.ID,
+	}
+
+	return m.DB.QueryRow(query, args...).Scan(&movie.Version)
 }
 
 // Delete movie by ID
 func (m MovieModel) Delete(id int64) error {
-	stmt := `
+	if id < 1 {
+		return ErrRecordNotFound
+	}
+
+	query := `
 		DELETE FROM movies
 		WHERE id = $1
 	`
-	_, err := m.DB.Exec(stmt, id)
-	return err
+
+	result, err := m.DB.Exec(query, id)
+	if err != nil {
+		return err
+	}
+
+	rowsAffected, err := result.RowsAffected()
+	if err != nil {
+		return err
+	}
+
+	if rowsAffected == 0 {
+		return ErrRecordNotFound
+	}
+	return nil
 }
